requirements.txt
absl-py==0.11.0
astunparse==1.6.3
awscli==1.18.209
botocore==1.19.49
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
easyocr==1.2.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
jmespath==0.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
mss==6.1.0
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python>=4.1.2.30
opt-einsum==3.3.0
pandas==1.2.0
picamera
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
scikit-image==0.18.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
tifffile==2020.12.8
torch==1.7.1
torchvision==0.8.2
tqdm==4.43.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.4.0
flask

tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1